InformationOrganisation/requirements.txt
asgiref==3.2.10
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.2
django-autocomplete-light==3.8.1
django-import-export==2.4.0
et-xmlfile==1.0.1
isodate==0.6.0
jdcal==1.4.1
MarkupPy==1.14
numpy==1.19.2
odfpy==1.4.1
openpyxl==3.0.5
pyparsing==2.4.7
pytz==2020.1
PyYAML==5.3.1
rdflib==5.0.0
six==1.15.0
SPARQLWrapper==1.8.5
sqlparse==0.4.1
tablib==2.0.0
whitenoise==5.2.0
xlrd==1.2.0
xlwt==1.3.0